+++ megalinter/reporters/ConsoleLinterReporter.py
@@ -49,8 +49,10 @@ class ConsoleLinterReporter(Reporter):
         base_phrase = f"Linted [{self.master.descriptor_id}] files with [{self.master.linter_name}]"
         elapse = str(round(self.master.elapsed_time_s, 2)) + "s"
         if self.master.return_code == 0 and self.master.status == "success":
-            logging.info(f"{base_phrase} successfully - ({elapse})")
+            logging.info(f"✅ {base_phrase} successfully - ({elapse})")
         elif self.master.return_code == 0 and self.master.status != "success":
-            logging.warning(f"{base_phrase}: Found non blocking error(s) - ({elapse})")
+            logging.warning(
+                f"✅ {base_phrase}: Found non blocking error(s) - ({elapse})"
+            )
         elif self.master.return_code != 0 and self.master.status != "success":
-            logging.error(f"{base_phrase}: Found error(s) - ({elapse})")
+            logging.error(f"❌ {base_phrase}: Found error(s) - ({elapse})")